Update main.sh
This PR removes the intermediate files sorted.txt, bssid.txt, and mac.txt that were being created and then immediately deleted. 
It also combines the sort and uniq commands into a single command. Finally, it removes the awk command since it is not necessary if uniq is used.
Also simplified some code parts here and there.